CBGroupsDao moved to templates cb package
